requirements.txt
django
djangorestframework
gunicorn 
psycopg2-binary
drf-yasg